+++ include/matx/operators/clone.h
@@ -119,9 +119,9 @@ namespace matx
     auto __MATX_INLINE__ clone(Op t, const std::array<index_t, Rank> &shape)
     {
       if constexpr (is_tensor_view_v<Op>) {
-        return t.template Clone<Rank>(shape);
+        return t.template Clone<static_cast<int>(Rank)>(shape);
       } else {
-        return detail::CloneOp<Rank, Op, index_t>(t, shape);
+        return detail::CloneOp<static_cast<int>(Rank), Op, index_t>(t, shape);
       }
     };
